[refactor] package 구조 수정, 컨트롤러 및 서비스 주석처리
