+++ src/main/java/com/honeyboard/api/project/finale/mapper/FinaleTeamMapper.java
@@ -1,45 +1,45 @@
-package com.honeyboard.api.project.finale.mapper;
-
-import java.time.LocalDate;
-import java.util.List;
-
-import com.honeyboard.api.user.model.UserName;
-import org.apache.ibatis.annotations.Param;
-
-public interface FinaleTeamMapper {
-
-	List<FinaleTeam> selectSubmitStatusByDate(LocalDate targetDate);
-
-	List<UserName> selectRemainedUsers(int generationId);
-
-	List<FinaleMember> selectTeamMembers(int teamId);
-
-	FinaleTeam selectTeamById(int teamId);
-
-	boolean existsTeamById(Integer teamId);
-
-	int deleteTeamMembers(Integer teamId);
-
-	int deleteTeamMember(
-			@Param("teamId") Integer teamId,
-			@Param("userId") Integer userId
-	);
-
-	int insertFinaleTeamWithMembers(
-			@Param("team") FinaleTeam team,
-			@Param("leaderId") Integer leaderId,
-			@Param("memberIds") List<Integer> memberIds
-	);
-
-	int updateTeamMembers(
-			@Param("teamId") Integer teamId,
-			@Param("leaderId") Integer leaderId,
-			@Param("memberIds") List<Integer> memberIds
-	);
-
-	List<Integer> findExistingTeamMembers(
-			@Param("userIds") List<Integer> userIds,
-			@Param("excludeTeamId") Integer excludeTeamId
-	);
-
-}
+//package com.honeyboard.api.project.finale.mapper;
+//
+//import java.time.LocalDate;
+//import java.util.List;
+//
+//import com.honeyboard.api.user.model.UserName;
+//import org.apache.ibatis.annotations.Param;
+//
+//public interface FinaleTeamMapper {
+//
+//	List<FinaleTeam> selectSubmitStatusByDate(LocalDate targetDate);
+//
+//	List<UserName> selectRemainedUsers(int generationId);
+//
+//	List<FinaleMember> selectTeamMembers(int teamId);
+//
+//	FinaleTeam selectTeamById(int teamId);
+//
+//	boolean existsTeamById(Integer teamId);
+//
+//	int deleteTeamMembers(Integer teamId);
+//
+//	int deleteTeamMember(
+//			@Param("teamId") Integer teamId,
+//			@Param("userId") Integer userId
+//	);
+//
+//	int insertFinaleTeamWithMembers(
+//			@Param("team") FinaleTeam team,
+//			@Param("leaderId") Integer leaderId,
+//			@Param("memberIds") List<Integer> memberIds
+//	);
+//
+//	int updateTeamMembers(
+//			@Param("teamId") Integer teamId,
+//			@Param("leaderId") Integer leaderId,
+//			@Param("memberIds") List<Integer> memberIds
+//	);
+//
+//	List<Integer> findExistingTeamMembers(
+//			@Param("userIds") List<Integer> userIds,
+//			@Param("excludeTeamId") Integer excludeTeamId
+//	);
+//
+//}
