+++ src/main/java/com/honeyboard/api/project/finale/model/Response/FinaleProjectBoardList.java
@@ -1,4 +1,4 @@
-package com.honeyboard.api.project.finale.model;
+package com.honeyboard.api.project.finale.model.Response;
 
 import lombok.*;
