+++ src/main/java/com/honeyboard/api/project/finale/service/FinaleProjectBoardService.java
@@ -1,19 +1,19 @@
-package com.honeyboard.api.project.finale.service;
-
-import java.util.List;
-
-public interface FinaleProjectBoardService {
-
-	List<FinaleProjectBoard> getAllFinaleBoards(int finaleId);
-
-	FinaleProjectBoard getFinaleBoard(int boardId);
-
-	boolean addFinaleBoard(FinaleProjectBoard board);
-
-	boolean updateFinaleBoard(int boardId, FinaleProjectBoard board);
-
-	void changeFinaleBoardCompletion(int boardId);
-
-	boolean softDeleteFinaleBoard(int boardId);
-
-}
+//package com.honeyboard.api.project.finale.service;
+//
+//import java.util.List;
+//
+//public interface FinaleProjectBoardService {
+//
+//	List<FinaleProjectBoard> getAllFinaleBoards(int finaleId);
+//
+//	FinaleProjectBoard getFinaleBoard(int boardId);
+//
+//	boolean addFinaleBoard(FinaleProjectBoard board);
+//
+//	boolean updateFinaleBoard(int boardId, FinaleProjectBoard board);
+//
+//	void changeFinaleBoardCompletion(int boardId);
+//
+//	boolean softDeleteFinaleBoard(int boardId);
+//
+//}
