+++ src/main/java/com/honeyboard/api/project/finale/service/FinaleProjectService.java
@@ -1,17 +1,17 @@
-package com.honeyboard.api.project.finale.service;
-
-import java.util.List;
-
-public interface FinaleProjectService {
-
-	List<FinaleProject> getAllFinaleProject(int generationId);
-
-	boolean addFinaleProjectThumbnail(String thumbnail);
-
-	boolean saveFinaleProject(FinaleProject finaleProject);
-
-	boolean updateFinaleProject(FinaleProject finaleProject);
-
-	boolean softDeleteFinaleProject(int teamId);
-
-}
+//package com.honeyboard.api.project.finale.service;
+//
+//import java.util.List;
+//
+//public interface FinaleProjectService {
+//
+//	List<FinaleProject> getAllFinaleProject(int generationId);
+//
+//	boolean addFinaleProjectThumbnail(String thumbnail);
+//
+//	boolean saveFinaleProject(FinaleProject finaleProject);
+//
+//	boolean updateFinaleProject(FinaleProject finaleProject);
+//
+//	boolean softDeleteFinaleProject(int teamId);
+//
+//}
